+++ src/bot.rs
@@ -23,6 +23,8 @@ pub fn framework_opts() -> FrameworkOptions<Data, anyhow::Error> {
             commands::music::join(),
             commands::music::disconnect(),
             commands::music::stop(),
+            commands::music::play(),
+            commands::music::queue(),
             commands::music::playing(),
             commands::music::dashboard(),
             commands::music::lyrics(),